Fix plaid cdn error, add confirm popup on delete pet

+++ www/pages/editPet/editPetService.js
@@ -1,6 +1,6 @@
 /* global angular */
 angular.module('editPetService', [])
-.factory('editPet', function ($http, $location, $rootScope) {
+.factory('editPet', function ($http, $location, $rootScope, $ionicPopup) {
   const edit = (pet) => {
     $http.put(`http://35.167.2.107:3000/v1/pets/${$rootScope.pet.id}`, pet)
       .then(() => {
@@ -11,11 +11,18 @@ angular.module('editPetService', [])
   };
 
   const deleter = () => {
-    return $http.delete(`http://35.167.2.107:3000/v1/pets/${$rootScope.pet.id}`)
-      .then(() => {
-        $location.path('/app/myPets');
-      }, (err) => {
-        console.warn(err);
+    $ionicPopup.confirm({
+      template: 'Are you sure you want to delete this pet?',
+    })
+      .then(res => {
+        if (res) {
+          return $http.delete(`http://35.167.2.107:3000/v1/pets/${$rootScope.pet.id}`)
+            .then(() => {
+              $location.path('/app/myPets');
+            }, (err) => {
+              console.warn(err);
+            });
+        }
       });
   };
   // you change pet details on this page
